requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
pdf2image==1.16.3
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.26.4
pytesseract==0.3.10
openai==1.3.7
python-dotenv==1.0.0
